modified delete endpoint

+++ app/food/food_route.py
@@ -35,8 +35,10 @@ async def get_all_food_ads(limit=10, skip=0):
 
 @router.delete("/{food_id}")
 async def delete_food_ad(food_id):
-
-    deleted_ad = await Food.delete_one(food_id)
+    if not food_id:
+        raise HTTPException(status.HTTP_400_BAD_REQUEST, "Food ID required")
+    deleted_ad = await Food.find(food_id)
     if not deleted_ad:
-        raise HTTPException(status.HTTP_404_NOT_FOUND, "No add found to delete!")
+        raise HTTPException(status.HTTP_404_NOT_FOUND, "Food not found")
+    await Food.delete_one(food_id)
     return {"message": "Ad successfully deleted!"}
